feat: add tabs to AddCustomListDialog
